+++ app/Livewire/Actions.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire;
+
+use Livewire\Component;
+
+class Actions extends Component
+{
+    public function render()
+    {
+        return view('livewire.actions');
+    }
+}