+++ resources/js/app.js
@@ -1,1 +1,2 @@
 import 'preline'
+// import '../../vendor/rappasoft/laravel-livewire-tables/resources/imports/laravel-livewire-tables-all.js';