+++ resources/views/components/dropdown.blade.php
@@ -1,6 +1,6 @@
 @props(['id', 'text'])
 <div>
-    <button id="{{$id}}" data-dropdown-toggle="dropdownDefaultCheckbox"
+    <button id="{{$id}}" data-dropdown-toggle="dropdown-request-filter-status"
         class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center inline-flex items-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
         type="button">{{ $text }} <svg class="w-2.5 h-2.5 ms-3" aria-hidden="true" xmlns="http://www.w3.org/2000/svg"
             fill="none" viewBox="0 0 10 6">
@@ -10,7 +10,7 @@
     </button>
 
     <!-- Dropdown menu -->
-    <div id="dropdownDefaultCheckbox"
+    <div id="dropdown-request-filter-status"
         class="z-10 hidden w-48 bg-white divide-y divide-gray-100 rounded-lg shadow dark:bg-gray-700 dark:divide-gray-600">
         <ul class="p-3 space-y-3 text-sm text-gray-700 dark:text-gray-200" aria-labelledby="{{$id}}">
             {{