+++ resources/views/components/widgets/btn-create-new.blade.php
@@ -1,0 +1,7 @@
+@props(['href' , "text" => "New"])
+
+<x-button status="primary" >
+    <a href="{{$href}}">
+        {{$text}}
+    </a>
+</x-button>